requirements.txt
vastai-sdk>=0.6.0
